==> django apps/is/data/2_add_recommending_books.ipynb <==
import json
import pandas as pd
import random
from sklearn.preprocessing import MultiLabelBinarizer

with open('./genres.json', 'r') as f:
    all_genres = json.load(f)
all_genres

mlb = MultiLabelBinarizer(classes=all_genres)
mlb.fit([all_genres])  # Fit on the full set of genres

import pickle
# Save the MultiLabelBinarizer to a file
with open('mlb_genres.pkl', 'wb') as f:
    pickle.dump(mlb, f)

with open('books_data.json', 'r') as f:
    books_data = json.load(f)

books_df = pd.DataFrame(books_data)
books_df.head()

# Load the MultiLabelBinarizer from the file
with open('mlb_genres.pkl', 'rb') as f:
    mlb_loaded = pickle.load(f)


# Encode genres for each book and update the genres_vector field
genres_vectors = []
for book in books_data:
    one_hot_encoded = mlb.transform([book['genres']])[0]
    book['genres_vector'] = one_hot_encoded.tolist()
    genres_vectors.append(one_hot_encoded)

from sklearn.neighbors import NearestNeighbors
# Fit KNN model on the genres_vector data
knn = NearestNeighbors(n_neighbors=16, metric='cosine')  # 16 to exclude the book itself
knn.fit(genres_vectors)

# Predict the 15 nearest neighbors for each book and add to recommended_books
for i, book in enumerate(books_data):
    distances, indices = knn.kneighbors([genres_vectors[i]], n_neighbors=16)
    nearest_ids = [books_data[idx]["title_complete"] for idx in indices[0] if idx != i][:15]  # Exclude itself
    book['recommended_books'] = nearest_ids
    book['price'] = random.randint(30, 100)

# Save the updated data with recommendations back to JSON
with open('books_data_with_recommendations.json', 'w') as f:
    json.dump(books_data, f, indent=4)
